requirements.txt
fastapi>=0.121.3
uvicorn[standard]>=0.32.1
apscheduler>=3.10.4
pandas>=2.2.0
python-jobspy>=0.5.0
supabase>=2.0.0
resend>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
numpy>=1.26.0

